+++ frontend/src/App.jsx
@@ -10,7 +10,7 @@ import Tutor from './components/Tutor'
 import Forum from './components/Forums/Forums'
 import NewPost from './components/Forums/NewPost'
 import PostPage from './components/Forums/PostPage'
-import Admin from './components/Admin'
+import Admin from './components/xx/Admin'
 import Dashboard from './components/Dashboard'
 import Account from './components/Account'
 import Profile from './components/Profile'
